requirements.txt
openai>=1.0
google-genai>=1.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
Pillow>=10.0
requests>=2.28
